% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{series.accession}
\alias{series.accession}
\title{series.accession}
\format{
A Large character with GSE:
\code{series.accession}
}
\usage{
series.accession
}
\description{
A dataset containing the GSE of GEOmirror
}
\keyword{datasets}
